# api/openapi.yaml
openapi: 3.1.0
info:
  title: LyraNest Community API
  version: 0.3.0
  description: Authenticated API for library discovery, playback, management, normalized sidecar lyrics, and persistent per-user listening data.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/tracks/{trackId}/stream:
    parameters:
      - name: trackId
        in: path
        required: true
        schema:
          type: string
    get:
      security: []
      operationId: streamTrack
      summary: Stream the original audio file
      description: Supports byte ranges for direct playback without transcoding.
      parameters:
        - name: Range
          in: header
          required: false
          description: Requested byte range, such as bytes=0-1023.
          schema:
            type: string
        - name: If-Range
          in: header
          required: false
          description: Strong ETag or HTTP-date validator for Range. A matching validator returns 206; a stale validator ignores Range and returns the complete 200 response.
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/AudioStream"
        "206":
          $ref: "#/components/responses/PartialAudioStream"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      security: []
      operationId: headTrackStream
      summary: Inspect the original audio stream
      responses:
        "200":
          $ref: "#/components/responses/AudioStreamHeaders"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
...
    NotFound:
      description: Track, audio file, artwork, or lyrics not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    AudioStream:
      description: Complete original audio file
      headers:
        Accept-Ranges:
          schema:
            type: string
          const: bytes
        ETag:
          schema:
            type: string
          description: Strong SHA-256 content validator in quoted hexadecimal form.
        Digest:
          schema:
            type: string
            example: sha-256=47DEQpj8HBSa+/TImW+5JCeuQeRkm5NMpJWZG3hSuFU=
          description: SHA-256 representation digest using the standard base64 form.
        X-Media-Version:
          schema:
            type: string
          description: Stable lowercase hexadecimal SHA-256 version of the complete audio representation.
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Last-Modified:
          schema:
            type: string
          description: File modification time in HTTP-date format.
        Content-Disposition:
          schema:
            type: string
          description: Inline disposition with the source file name.
      content:
        audio/mpeg:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/flac:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/mp4:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/ogg:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/wav:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/AudioBinary"
    PartialAudioStream:
      description: Requested byte range from the original audio file
      headers:
        Accept-Ranges:
          schema:
            type: string
          const: bytes
        ETag:
          schema:
            type: string
          description: Strong SHA-256 content validator in quoted hexadecimal form.
        Digest:
          schema:
            type: string
            example: sha-256=47DEQpj8HBSa+/TImW+5JCeuQeRkm5NMpJWZG3hSuFU=
          description: SHA-256 representation digest using the standard base64 form.
        X-Media-Version:
          schema:
            type: string
          description: Stable lowercase hexadecimal SHA-256 version of the complete audio representation.
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Content-Range:
          schema:
            type: string
          description: Byte range included in this response.
        Last-Modified:
          schema:
            type: string
          description: File modification time in HTTP-date format.
        Content-Disposition:
          schema:
            type: string
          description: Inline disposition with the source file name.
      content:
        audio/mpeg:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/flac:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/mp4:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/ogg:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/wav:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/AudioBinary"
    AudioStreamHeaders:
      description: Original audio file headers with no response body
      headers:
        Accept-Ranges:
          schema:
            type: string
          const: bytes
        ETag:
          schema:
            type: string
          description: Strong SHA-256 content validator in quoted hexadecimal form.
        Digest:
          schema:
            type: string
            example: sha-256=47DEQpj8HBSa+/TImW+5JCeuQeRkm5NMpJWZG3hSuFU=
          description: SHA-256 representation digest using the standard base64 form.
        X-Media-Version:
          schema:
            type: string
          description: Stable lowercase hexadecimal SHA-256 version of the complete audio representation.
        Content-Type:
          schema:
            type: string
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Last-Modified:
          schema:
            type: string
          description: File modification time in HTTP-date format.
        Content-Disposition:
          schema:
            type: string
          description: Inline disposition with the source file name.
  schemas:
...
    AudioBinary:
      type: string
      format: binary
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string